project-local GenericTypeDefs.h deprecated; replaced by $SPJ_SWR_LOC)/libs_common/public/spj_stdint.h
.project:
    - Remove folder /GenericTypes
.cproject
    - now include $SPJ_SWR_LOC)/libs_common/public
bcd/bcd_bin.c:
    - Fix bug in _u32toBCD()
legal_YMDHMS.c:
    - Remove unused #pragma GCC
to_sys_endian_u32.c, reverse_u32.c:
    - fix promotion error in reverse_u32.c
c8_bag.c, u8_bag.c, heap_log.c, secs_to_hms32_str.c:
    - fix printf/sprint long/unsigned long cast errors.
heap_1way_u16.c, tbl_search_Ascending_key_u16.c:
    - Add 'uintptr_t' to fix incorrect arithmetic on void*
read_next_number.c, read_next_float.c:
    - Avoid unwanted double arithmetic (float only)
misc files:
    - fix singled-to-unsigned char pointer warnings (flagged in CL430) build.

+++ src/stores/c8_bag.c
@@ -130,7 +130,7 @@ PUBLIC void C8bag_Print(S_C8bag const *m)
    if(sizeof(m->lines[0]) == sizeof(U32)) {
       U8 i;
       for(i = 0; i < _C8bag_NumLines; i++) {
-         printf("  %d: 0x08%lX\r\n", i, m->lines[i]); }}
+         printf("  %d: 0x08%lX\r\n", i, (unsigned long)m->lines[i]); }}
 
 }
 
@@ -149,7 +149,7 @@ PUBLIC C8 const * C8bag_PrintLine(C8 *out, S_C8bag const *m)
                : (i == _C8bag_NumLines-1
                   ? " %08lX]"                 // End with right-bracket.
                   : " %08lX"),
-            m->lines[i]); }
+            (unsigned long)m->lines[i]); }
       return out; }
    return strcpy(out, "[C8Bag print error]");
 }
